Extract and clean up command impl
